management/requirements.txt
grpcio
grpcio-tools
mysql-connector-python
Flask
kafka-python